# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TalentLock API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /freelancers:
    get:
      operationId: listFreelancers
      tags: [freelancers]
      summary: List freelancer profiles
      parameters:
        - name: field
          in: query
          schema:
            type: string
          description: Filter by field of work
        - name: available
          in: query
          schema:
            type: boolean
          description: Filter by availability
        - name: verified
          in: query
          schema:
            type: boolean
          description: When true, only freelancers with at least one verified document
        - name: limit
          in: query
          schema:
            type: integer
          description: Number of results
        - name: offset
          in: query
          schema:
            type: integer
          description: Pagination offset
        - name: availableFrom
          in: query
          schema:
            type: string
            format: date
          description: Only return freelancers available on or before this date
        - name: q
          in: query
          schema:
            type: string
          description: Full-text keyword search across bio and skills
        - name: professionCategory
          in: query
          schema:
            type: string
            enum: [technology, education]
          description: Filter by profession category
        - name: teachingSubject
          in: query
          schema:
            type: string
          description: Case-insensitive substring match on teachingSubjects (education professionals)
      responses:
        "200":
          description: List of freelancers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FreelancerProfile"
    post:
      operationId: createFreelancerProfile
      tags: [freelancers]
      summary: Create freelancer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFreelancerProfileBody"
      responses:
        "201":
          description: Created freelancer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerProfile"
  /freelancers/me:
    get:
      operationId: getMyFreelancerProfile
      tags: [freelancers]
      summary: Get my freelancer profile
      responses:
        "200":
          description: My freelancer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateMyFreelancerProfile
      tags: [freelancers]
      summary: Update my freelancer profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFreelancerProfileBody"
      responses:
        "200":
          description: Updated freelancer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerProfile"
  /freelancers/{id}:
    get:
      operationId: getFreelancerProfile
      tags: [freelancers]
      summary: Get a freelancer profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Freelancer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerProfileDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── User ──────────────────────────────────────────────────────────────────
...
    ProfessionCategory:
      type: string
      enum: [technology, education]
    EducationProfessionType:
      type: string
      enum: [school_teacher, university_lecturer, tutor, researcher]
    HighestDegree:
      type: string
      enum: [bachelors, masters, phd, postdoc]
    DbsCheckStatus:
      type: string
      enum: [not_uploaded, uploaded, verified, expired]
    PreferredTeachingMode:
      type: string
      enum: [in_person, online, both]
...
    FreelancerProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        clerkId:
          type: string
        name:
          type: string
        tagline:
          type: string
        fieldOfWork:
          type: string
        skills:
          type: array
          items:
            type: string
        yearsExperience:
          type: integer
        bio:
          type: ["string", "null"]
        resumeUrl:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        paymentPreference:
          type: string
          description: "hourly or daily"
        hourlyRate:
          type: ["number", "null"]
        dailyRate:
          type: ["number", "null"]
        achievements:
          type: ["string", "null"]
        isVerified:
          type: boolean
        verificationLevel:
          type: string
          enum: [unverified, partially_verified, fully_verified]
        verificationStatus:
          type: string
          description: "unverified | pending | verified | rejected"
        verificationNote:
          type: ["string", "null"]
        documentNames:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        currentBookingId:
          type: ["integer", "null"]
        bookingEndDate:
          type: ["string", "null"]
          format: date-time
        subscriptionPlan:
          type: string
        availableFrom:
          type: ["string", "null"]
          format: date-time
        nextAvailableDate:
          type: ["string", "null"]
          format: date
          description: Cached next available date (null when not available)
        availabilityNote:
          type: ["string", "null"]
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        completenessScore:
          type: integer
          description: Profile completeness score 0–100 (Talent Vault requires ≥ 60)
        professionCategory:
          $ref: "#/components/schemas/ProfessionCategory"
        educationProfessionType:
          allOf:
            - $ref: "#/components/schemas/EducationProfessionType"
          nullable: true
        teachingSubjects:
          type: ["array", "null"]
          items:
            type: string
        teachingLevels:
          type: ["array", "null"]
          items:
            type: string
        yearsTeachingExperience:
          type: ["integer", "null"]
        highestDegree:
          allOf:
            - $ref: "#/components/schemas/HighestDegree"
          nullable: true
        degreeSubject:
          type: ["string", "null"]
        degreeInstitution:
          type: ["string", "null"]
        teachingLicenceState:
          type: ["string", "null"]
        teachingLicenceExpiry:
          type: ["string", "null"]
          format: date-time
        dbsCheckStatus:
          allOf:
            - $ref: "#/components/schemas/DbsCheckStatus"
          nullable: true
        researchPublications:
          type: ["string", "null"]
        preferredTeachingMode:
          allOf:
            - $ref: "#/components/schemas/PreferredTeachingMode"
          nullable: true
        location:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - clerkId
        - name
        - tagline
        - fieldOfWork
        - skills
        - yearsExperience
        - paymentPreference
        - isVerified
        - verificationLevel
        - isAvailable
        - subscriptionPlan
        - reviewCount
        - completenessScore
        - professionCategory
        - createdAt
...
    FreelancerVerification:
      type: object
      properties:
        level:
          type: string
          description: "unverified | partially_verified | fully_verified"
        verifiedDocumentCount:
          type: integer
      required:
        - level
        - verifiedDocumentCount
    FreelancerProfileDetail:
      allOf:
        - $ref: "#/components/schemas/FreelancerProfile"
        - type: object
          properties:
            email:
              type: ["string", "null"]
            verification:
              $ref: "#/components/schemas/FreelancerVerification"
          required:
            - verification
    CreateFreelancerProfileBody:
      type: object
      properties:
        tagline:
          type: string
        fieldOfWork:
          type: string
        skills:
          type: array
          items:
            type: string
        yearsExperience:
          type: integer
        resumeUrl:
          type: ["string", "null"]
        paymentPreference:
          type: string
        hourlyRate:
          type: ["number", "null"]
        dailyRate:
          type: ["number", "null"]
        achievements:
          type: ["string", "null"]
        subscriptionPlan:
          type: string
        resumeAnalysis:
          $ref: "#/components/schemas/ResumeAnalysis"
          nullable: true
        professionCategory:
          $ref: "#/components/schemas/ProfessionCategory"
        educationProfessionType:
          allOf:
            - $ref: "#/components/schemas/EducationProfessionType"
          nullable: true
        teachingSubjects:
          type: ["array", "null"]
          items:
            type: string
        teachingLevels:
          type: ["array", "null"]
          items:
            type: string
        yearsTeachingExperience:
          type: ["integer", "null"]
        highestDegree:
          allOf:
            - $ref: "#/components/schemas/HighestDegree"
          nullable: true
        degreeSubject:
          type: ["string", "null"]
        degreeInstitution:
          type: ["string", "null"]
        teachingLicenceState:
          type: ["string", "null"]
        teachingLicenceExpiry:
          type: ["string", "null"]
          format: date-time
        dbsCheckStatus:
          allOf:
            - $ref: "#/components/schemas/DbsCheckStatus"
          nullable: true
        researchPublications:
          type: ["string", "null"]
        preferredTeachingMode:
          allOf:
            - $ref: "#/components/schemas/PreferredTeachingMode"
          nullable: true
        location:
          type: ["string", "null"]
      required:
        - tagline
        - fieldOfWork
        - skills
        - yearsExperience
        - paymentPreference
        - subscriptionPlan
...
    UpdateFreelancerProfileBody:
      type: object
      properties:
        tagline:
          type: string
        bio:
          type: ["string", "null"]
        fieldOfWork:
          type: string
        skills:
          type: array
          items:
            type: string
        yearsExperience:
          type: integer
        resumeUrl:
          type: ["string", "null"]
        portfolioUrl:
          type: ["string", "null"]
        paymentPreference:
          type: string
        hourlyRate:
          type: ["number", "null"]
        dailyRate:
          type: ["number", "null"]
        achievements:
          type: ["string", "null"]
        subscriptionPlan:
          type: string
        isAvailable:
          type: boolean
        availableFrom:
          type: ["string", "null"]
          format: date-time
        availabilityNote:
          type: ["string", "null"]
        professionCategory:
          $ref: "#/components/schemas/ProfessionCategory"
        educationProfessionType:
          allOf:
            - $ref: "#/components/schemas/EducationProfessionType"
          nullable: true
        teachingSubjects:
          type: ["array", "null"]
          items:
            type: string
        teachingLevels:
          type: ["array", "null"]
          items:
            type: string
        yearsTeachingExperience:
          type: ["integer", "null"]
        highestDegree:
          allOf:
            - $ref: "#/components/schemas/HighestDegree"
          nullable: true
        degreeSubject:
          type: ["string", "null"]
        degreeInstitution:
          type: ["string", "null"]
        teachingLicenceState:
          type: ["string", "null"]
        teachingLicenceExpiry:
          type: ["string", "null"]
          format: date-time
        dbsCheckStatus:
          allOf:
            - $ref: "#/components/schemas/DbsCheckStatus"
          nullable: true
        researchPublications:
          type: ["string", "null"]
        preferredTeachingMode:
          allOf:
            - $ref: "#/components/schemas/PreferredTeachingMode"
          nullable: true
        location:
          type: ["string", "null"]

    # ── Employer Profile ───────────────────────────────────────────────────────
...
    ResumeWorkExperience:
      type: object
      properties:
        company:
          type: string
        role:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        highlights:
          type: array
          items:
            type: string
      required: [company, role, startDate, endDate, highlights]
    ResumeEducation:
      type: object
      properties:
        institution:
          type: string
        degree:
          type: string
        year:
          type: string
      required: [institution, degree, year]
    ResumeAnalysis:
      type: object
      properties:
        workExperience:
          type: array
          items:
            $ref: "#/components/schemas/ResumeWorkExperience"
        education:
          type: array
          items:
            $ref: "#/components/schemas/ResumeEducation"
        certifications:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
      required: [workExperience, education, certifications, languages]

    # ── Team (Enterprise) ──────────────────────────────────────────────────────
...